# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizApp API Documentation",
        "contact": {
            "email": "haykalcs@gmail.com"
        },
        "version": "1"
    },
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get List of Students",
                "description": "Get List of Students",
                "operationId": "getStudentsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data list siswa"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "id",
                                                    "name",
                                                    "email"
                                                ],
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    },
                                                    "avatar": {
                                                        "type": "string"
                                                    },
                                                    "number": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 2,
                                                    "name": "Alex",
                                                    "email": "alex@gmail.com",
                                                    "avatar": null,
                                                    "number": null
                                                },
                                                {
                                                    "id": 3,
                                                    "name": "Bill",
                                                    "email": "bill@gmail.com",
                                                    "avatar": "1850228838.jpg",
                                                    "number": 1
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get List of All Users",
                "description": "Get List of All Users",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Seluruh user"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "id",
                                                    "name",
                                                    "email",
                                                    "role",
                                                    "created_at",
                                                    "updated_at"
                                                ],
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "role": {
                                                        "type": "string"
                                                    },
                                                    "avatar": {
                                                        "type": "string"
                                                    },
                                                    "number": {
                                                        "type": "integer"
                                                    },
                                                    "last_seen": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 2,
                                                    "name": "Kira",
                                                    "email": "kira@gmail.com",
                                                    "email_verified_at": "2021-09-15T08:14:28.000000Z",
                                                    "role": "guru",
                                                    "avatar": null,
                                                    "number": null,
                                                    "last_seen": "2021-09-20 02:56:55",
                                                    "created_at": "2021-09-15T08:14:28.000000Z",
                                                    "updated_at": "2021-09-20T02:56:55.000000Z"
                                                },
                                                {
                                                    "id": 3,
                                                    "name": "John Doe",
                                                    "email": "johndoe@gmail.com",
                                                    "email_verified_at": "2021-09-15T08:14:28.000000Z",
                                                    "role": "siswa",
                                                    "avatar": "1850228838.jpg",
                                                    "number": 1,
                                                    "last_seen": "2021-09-20 02:56:55",
                                                    "created_at": "2021-09-15T08:14:28.000000Z",
                                                    "updated_at": "2021-09-20T02:56:55.000000Z"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/teachers": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get List of Teachers",
                "description": "Get List of Teachers",
                "operationId": "getTeachersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data list guru"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "id",
                                                    "name",
                                                    "email"
                                                ],
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    },
                                                    "avatar": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 2,
                                                    "name": "Kira",
                                                    "email": "kira@gmail.com",
                                                    "avatar": null
                                                },
                                                {
                                                    "id": 3,
                                                    "name": "Mita",
                                                    "email": "mita@gmail.com",
                                                    "avatar": "1850228838.jpg"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User Detail",
                "description": "Get User Detail",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data user"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 3,
                                                "name": "John Doe",
                                                "email": "johndoe@gmail.com",
                                                "email_verified_at": "2021-09-15T08:14:28.000000Z",
                                                "role": "siswa",
                                                "avatar": "1850228838.jpg",
                                                "number": 1,
                                                "last_seen": "2021-09-20 02:56:55",
                                                "created_at": "2021-09-15T08:14:28.000000Z",
                                                "updated_at": "2021-09-20T02:56:55.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}